+++ tests/PinguApps.Appwrite.Server.Tests/Servers/Account/AccountServerTests.cs
@@ -25,18 +25,3 @@ public partial class AccountServerTests
         _appwriteServer = serviceProvider.GetRequiredService<IAppwriteServer>();
     }
 }
-
-public static class AccountTestsExtensions
-{
-    public static MockedRequest ExpectedHeaders(this MockedRequest request)
-    {
-        return request
-            .WithHeaders("x-appwrite-project", Constants.ProjectId)
-            .WithHeaders("x-appwrite-key", Constants.ApiKey)
-            .WithHeaders("x-sdk-name", Constants.SdkName)
-            .WithHeaders("x-sdk-platform", "server")
-            .WithHeaders("x-sdk-language", Constants.SdkLanguage)
-            .WithHeaders("x-sdk-version", Constants.SdkVersion)
-            .WithHeaders("x-appwrite-response-format", Constants.AppwriteResponseFormat);
-    }
-}
